msm: pil-q6v4-lpass: Combine SSR and PIL drivers
Combine both drivers into one so that we can merge the PIL and
SSR frameworks. To cope with the differences between 8960 and
9615 (the former supports PIL while the latter does not) we check
to see if there is any platform data specified for the device. If
there's platform data we know we're driving a PIL supported
device and act accordingly. On 9615, we only support calling
panic on lpass watchdog interrupts, so it's safe to avoid
ioremapping registers.

Change-Id: I03ffef481c6c6cf4a30d00b2cff8df65830d2aaf
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>