msm: pil: Add memory map tracking logic
Track the memory areas for each image that PIL loads. This
information is useful for ramdump collection and general
debugging.

Also, push the struct throughout the code to remove the
dependence of the segment loading code on the elf headers. This
paves the way for adding support for relocatable images without
having to touch the segment loading code.

Change-Id: I0cc2b24c2cf89dd176726f87ea08c082e3ff893a
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>