msm: pil-venus: Implement Secure PIL support
Use scm-pas to detect if Secure PIL is required and if it is,
call scm-pas APIs to boot and shutdown the Venus Processor
instead of handling these procedures in the Linux Kernel.

Also, Linux Kernel is responsible for controlling GDSC and
clocks outside of Venus Subsystem in the Secure PIL case.
Therefore, wrap up scm-pas APIs with necessary codes to control
these resources.

Change-Id: Id04f2694072e13e03548807bbc92e983378744c0
Signed-off-by: Tianyi Gou <tgou@codeaurora.org>