msm: ocmem: Add support for locking and unlocking regions
Add the common infrastucture required for locking and unlocking
OCMEM regions. This works with both the secure processor (TZ) mode
and in non-secure mode.

In non-secure mode the OCMEM driver is responsible for setting up
the region modes, the graphics OCMEM MPU and clearing OCMEM
regions using the DM clearing engine.

Change-Id: Iaf7af3977a0a581cddd42bd7fe5f0f476bf3aba0
Signed-off-by: Naveen Ramaraj <nramaraj@codeaurora.org>