msm: pil-q6v5-mss: Temporarily program MSS_MPLL1_CONFIG_CTL
Early samples of 8974 v1.0 hardware misconfigure the MPLL1 in
the Modem's primary bootloader. Since this ROM code cannot be
changed, work around the issue by setting a corrected configuration
register value before bringing the modem out of reset.

This change will be reverted when modem support on v1.0 hardware is
dropped, which is expected in the very near future.

Change-Id: Ia422e97730c48b5fd6a6fae084eafd9aa0ee34de
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>