adiag: Add new entry in message mask table
DIAG driver maintains a table of F3 masks. This table
has to be in sync with modem mask information. Modem
updates are not synchronized with Linux code updates.
Hence, sometimes modem and Linux go out of sync. This
results in missing F3 messages.

Adding this entry will resolve the dependency of manual
syncing and will make mask updates dynamic.

CRs-Fixed: 388505
Change-Id: I23fbe7171bb33cc0723d1b0cb069b41d62ba4231
Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>