usb: gadget: fix configuration descriptor for super speed devices
The bMaxPower field in the configuration descriptor should be specified
in different units for super speed devices. This field units are 2mA for
high speed connection and 8mA for super speed connection.

Without this fix, USB3 SS host drivers refuse to configure the super speed
device since it request for too much current to draw.

Change-Id: I586b92940163089ae691b2bb627712a723e68e36
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>