msm: clock-rpm: Set "last set khz" field to non-zero value during handoff
When a RPM (RPM SMD) clock is handed off, the last set KHz field is not
updated to a non-zero value. This means that despite the clock ref count
being non-zero, a vote will never be sent to RPM till the first non-zero
clk_set_rate() call is made on the clock.

If this call doesn't happen before late init, it's possible that the RPM
will turn off the clock when the application processor is using it. In a
real world scenario, we always get a non-zero clk_set_rate() call before
late init. So, this doesn't cause any real issues. Just fixing it for
completeness sake.

Change-Id: I48f43976e8a64da5060c16fd27b6b77d537b118c
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>