power: qpnp-charger: fix USBIN_VALID_IRQ
Fix a mixup in the device tree configuration which
caused qpnp-charger to request the coarse_detect
interrupt instead of the usbin_valid interrupt.

With the order of interrupts correctly restored,
there is a possibilty that the existing hack
calls the usbin_valid interrupt handler in error
and cause USB enumeration to fail.

Therefore, remove the hack which was used as a
workaround for the incorrect interrupt behaviour in the
qpnp-charger driver.

Change-Id: I170aa4d5f4739f12870bf18edc0e7db4841291bd
Signed-off-by: David Keitel <dkeitel@codeaurora.org>