power: msm_battery: Make BATTERY_MSM default n
When CONFIG_POWER_SUPPLY is enabled, CONFIG_BATTERY_MSM ends up
getting pulled in by default as a module. This is unnecessary
and requires an additional entry in defconfig files to disable
it. Additionally this has link-time dependencies provided by
CONFIG_MSM_ONCRPCROUTER so add that as a "depends on" as well.

Change-Id: Ie2aed6c66961e085c8b96e51b6bc5908d0ef32dd
Signed-off-by: Jack Pham <jackp@codeaurora.org>