usb: gadget: Prevent queueing of more than one request to host
Only one notification to a host may be queued at each period of time.
To ensure that, queue notification only after the completion
of previous notification.
If two QBI messages arrive to USB in a short period of time,
only one of the notifications may be sent to host immediately.
The second notification is sent to host from the completion callback
of first notification.

CRs-Fixed: 406652
Change-Id: I359ef567a1d704af115111dcc49fdf7e0e2cce19
Signed-off-by: Anna Perel <aperel@codeaurora.org>