arm/dt: msm8974: add MHL SII 8334 to device tree
This change includes the msm8974 FLUID DTS file changes
to add SII 8334 to the device tree, since SII 8334 is included
only with FLUID boards in 8974 chipsets. The MHL-SII8334 Tx
is connected to MSM over I2C via BLSP11 bus so we add a DTSI
config block as slave device under the addr corresponding to that
of BLSP11 bus.

Change-Id: I47be91cd9b17063141e2118ff87b942dfa7a5ecc
Signed-off-by: Manoj Rao <manojraj@codeaurora.org>