gpio: qpnp-pin: Make source select property name consistent
Code fragments and comments currently make reference to 'select'
instead of the documented name 'sec-select'. Change this to be
consistent with respect to each other for clarity.

Since we're modifying this anyways, let's pick a more sensible
name. To be consistent with the scheme used by the 'vin-sel'
property, let's change the source select property to 'src-sel'.
This requires modifying the code and updating all existing Device
Tree entries currently using this property.

Change-Id: Ib152340337a72dc744801e8a42ed4ea152565d0d
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>