msm: mdss: replace panel function pointers with single event handler
There are many events in MDP that are important to acommodate the needs
of different panel drivers, currently these events are propagated with
function pointers. However as more use cases come into picture the
ammount of events has increased, and not all of these are handled by all
panel drivers. In order to provide all this information to panel driver
and be able to extend it in future, create a generic way to propagate
events from MDP to panel driver through a single event handler.

Change-Id: Ide243f08bc4fa966ef9223b0a665cff95b9529db
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>