power: smb137c-charger: Correct USB OTG mode enable logic
Modify the logic that is used when a USB power supply change is
detected so that the SMB137C charger chip enters USB OTG mode
based exclusively on the value of the SCOPE power supply
framework property.  The value of the ONLINE property should
only be considered in order to enable or disable charging when
not in USB OTG mode.

Also remove an unnecessary TODO comment.

Change-Id: Ie53c4fcd7d3c6a18caeab96ec48f88906cfda0ee
Signed-off-by: David Collins <collinsd@codeaurora.org>