msm: pil: Store start address and size of images into IMEM
With relocatable image support there is no easy way to determine
from a ramdump where a particular image was loaded into DDR
without carefully parsing the kernel image and chasing a bunch of
pointers around. Simplify this task by writing the name of the
image and the start address and size of the image to IMEM.

Each record takes up exactly 20 bytes and there are only 10
records allocated. To ensure that we don't overwrite other
records, fail pil_desc_init() if we've already allocated all
available records.

Change-Id: I1e76a30d83bfc97ba418cdab96aa6410884d561b
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>