diag: Reorganize/modularize diag smd implementation
The diag smd implementation has many instances where code
is duplicated based upon the peripheral and the type of smd
channel that it works with. Also, addition of smd channels
will only compound this duplication. Smd code should be
reorganized to remove this duplicated code. Modularization of
the diag smd code should be used to assist in removal of duplicated
code and ease addition of new smd channel peripherals and
channel types in the future.

Change-Id: Ie0fa9474f646bd0bfa5c9dbc5d95007f315be596
Signed-off-by: Dixon Peterson <dixonp@codeaurora.org>