msm: pil: Migrate to SMP2P framework from SMSM mechanism.
State information is communicated with the subsystems via SMSM
bits that use a broadcast protocol. There are disadvantages
to this, such as, not being able to ask one subsystem to stop
without affecting the others. The new SMSM Point-to-Point (SMP2P)
framework allows point to point communication of state information
between subsystems. So, switch over to using SMP2P framework for
communicating low level state between the Application processor
and other subsystems. The SMP2P framework exposes the functionality
as standard GPIO and IRQ APIs. So, we just treat it as such.

Change-Id: I4c7d4be15ad594f5eb1e775266881f7e11047f5f
Signed-off-by: Seemanta Dutta <seemanta@codeaurora.org>