ASoC: wcd9320: Add workaround to prevent leakage on the spkdrv
The wcd9320 codec hardware has an issue that it can ends up having leakage
on the speaker drive.  This leakage can be avoided by keeping speaker
drive's PA and bandgap on.

Add and enable the workaround that turns PA and bandgap on by default with
a kernel parameter to on/off.

Change-Id: I6b607691b11260ddea01380201fa90f0da8cccf8
Signed-off-by: Joonwoo Park <joonwoop@codeaurora.org>