mmc: msm_sdcc: don't perform busy wait for CMD13 (status command)
Currently SDCC driver is waiting for the PROG_DONE interrupt after
sending out the CMD13 to card  even if command flags doesn't indicate
so. This may cause the issue in when we try to HPI (High Priority
Interrupt) the ongoing BKOPs (Background Operations). Hence this
change removes the busy wait (PROG_DONE) for CMD13.

Change-Id: I8fb6ab4908bf1c4c5a939413d27eec29cde22f68
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>