power: qpnp-bms: remember the temperature when ocv was taken
The code looks up the percent charge based on ocv at the
current temperature. The ocv could have been recorded earlier
under different temperature conditions. Looking it up at the
current temperature is wrong.

Hence remember the temperature when the ocv was first seen. And
use that to lookup the ocv values.

Change-Id: I416cb367f5d73421ef805952585aadb961014492
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>