slim_msm: Fix QMI message lengths
The QMI request and response descriptors have a maximum message length
field that is incorrectly set.  The length needs to be specified as the
expected maximum length of the QMI messages once it is encoded and not
the length of the request or response structure.  Specifying the
incorrect length could result in data corruption.

CRs-Fixed: 438809
Change-Id: I14aff6f823f0ba0685ee5678b99746364a8021db
Signed-off-by: Kenneth Heitke <kheitke@codeaurora.org>