USB: u_ether: Check if driver_data and port_usb is NULL before accessing
Rndis tx requests are queued to the controller, while disconnect is
in progress. There is a chance of getting the completion of those requests
after ep->driver_data and dev->port_usb are set to NULL.
This leads to NULL pointer dereference crash on tx_complete().

Hence, check if ep->driver_data and dev->port_usb are NULL before
accessing them in tx_complete() and free the current request.

CRs-Fixed: 435636
Change-Id: I42a427f7e51ffe4f81768756d835d123778df260
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>