msm: smp2p: Removing locking during interrupt notification
If an inbound entry changes, both the inbound and outbound entry locks
are held while performing an interrupt notification.  If a client then
attempts to set a new outbound bit in the notification function, then a
deadlock will occur.

Verify that the outbound entry is open (which is a terminal state) and
then unlock the outbound spinlock before doing the notification.

Change-Id: I662b721ef5fd48e75d12019bf12e370f4a7eea9d
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>