power: qpnp-bms: estimate OCV when a new battery is inserted
When a new battery is inserted, the BMS will continue to use its old
OCV value until a new OCV is taken. This is clearly wrong, as the new
battery can have a completely different state of charge than the old
one.
Fix this by making BMS driver estimate a new OCV based on the close
circuit battery voltage and IR drop across the battery when a battery
insertion is detected.

Change-Id: Ia0bdbb3d86a8698e1804e44fdfbda4ab4cf66fef
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>