msm: camera: Fix camera recording corruption
For ISP write master configuration, the buffer height
should be configured as image height instead of buffer
scanlines because for cases where scanlines is larger than
image height, bottom padding can be ignored by
not writing to the buffer. In current configuration,
image data of the next frame will be written to padding,
which result in image corruption.

Change-Id: I42b30248a97b485932b3c2af5f2cf697833a8a7a
Signed-off-by: Kevin Chan <ktchan@codeaurora.org>