msm: audio: qdsp6v2: Check return code from audio_aio_open()
When audio_aio_open() is triggered in blocking fashion, it  frees audio
resources and returns error code. In the current implementation,  callers
did not check the error code which can lead to dereference to freed
resources and cause system crash.  Update the callers to check return
code and then return error back to application to avoid the issue. In
the meanwhile, update the audio_aio_open not to free memory for caller
and leave the caller to free the memory in error condition.

Change-Id: I431aed57eb08154bcb896fd3e38dfbd2257f8742
CRs-fixed: 436653
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>