iommu: msm: Pass correct handle to interrupt
The IOMMU page fault handler expects the cookie passed to be of
struct msm_iommu_ctx_drvdata. However, the code was passing the
cookie as a struct platform_device pointer.

Change the registration of the page fault interrupt to pass in a
cookie corresponding to the msm_iommu_ctx_drvdata for the context
bank.

Change-Id: I713a81531e73207ed6cebf35bbb568688c438cd9
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>