usb: wwan: Fix RX data corruption issue
The received data is pushed to TTY in global workqueue context.
The work is scheduled from URB completion handler and TTY unthrottle
callback.  The global workqueue does not guarantee non-reentrance.
The data is getting corrupted when work is run simultaneously on
multiple processors.  Use non-reentrant version of global workqueue.

CRs-Fixed: 458325
Change-Id: I9562fc5be9d15c34cc4498fa3ddb325b35136fd2
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>