msm: rpm-regulator-smd: Add voltage and current forced update properties
Add support for qcom,always-send-voltage and
qcom,always-send-current rpm-regulator-smd device tree
properties.  These properties can be specified for a given
regulator in order to force updates to RPM voltage and current
parameters.  The default behavior is to only send updated voltage
and current parameter values to the RPM if a Linux consumer has
enabled the regulator.

Change-Id: I6ab02a7939229a515e8ad499aed5637a0f213532
Signed-off-by: David Collins <collinsd@codeaurora.org>