gpio: msm: Create an irq mapping if it doesn't exist for gpio_to_irq
Currently, the gpio-irqs are allocated through the device tree
interrupts property. A mapping will not exist for a gpio which
is not specified in the interrupts property. Thus, gpio_to_irq
will return a valid irq number for a gpio only if that gpio has
been specified as an interrupt in device tree.
To make gpio_to_irq work for all irqs, use the irq_create_mapping
function to create a mapping if one does not already exist.

Change-Id: I70ba2daa9a02f6e612676c0657919f33a9a72e9f
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>
Signed-off-by: Amy Maloche <amaloche@codeaurora.org>