thermal: tsens: Add local initialization flag
TSENS control registers are now configured at boot. Therefore
there is no need to re-initialize them now. The control
registers are also in a secure domain. Add a flag to check if
local initialization is required if the control registers
initialization is not done at boot. The change also enables
critical temperature thresholds that shut down the device
if the temperature crosses 120degC/-35degC to protect
the device from overheating. Also update the polling rate
default from 312.5ms to 62.5ms for cases when local control
register initialization is done.

Change-Id: I73ded3d1656a0fd168c9e8d742bccf3226c162cf
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>