msm: vidc: Initialize frequency to the highest value
Previously the driver was initializing the frequency to the lowest value
then scaling it up according to the load.  However, if the load is
beyond the max load in the load_table, the loop will quit without
setting a proper frequency.  Hence, it's better for us to initialize
with the max supported frequency.

Change-Id: I969afde37e03ffb49ceb36dc96b65ab9b161c7f4
CRs-Fixed: 463761
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>