power: qpnp-bms: update correction algorithm
Correct only when the soc estimate is lower than 15. This way we won't
affect linearity and accuracy for most part. The corrections will
come in only when the soc est gets low.

Also repurpose the device tree adjust soc low threshold property so that
it determines the point the estimated SoC needs to be at to start
adjusting. Remove the high threshold since it is unused.

Change-Id: I47f909aa75d96d4f7735e0d17521e4a6b37df914
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>