ASoC: msm: Pull down mic bias output by default
When mic bias is high Z by default and LDOH is enabled,
the mic bias output can pull-up to a non-zero voltage if there is
no loading or the load leakage is very small. But pulling down
mic bias output is not in high Z to eliminate the floating voltage.
And this change bring about sleep current increase because micbias4
try Current SRC. So, this current SCR for micbais4 is disabled.

Change-Id: I3acc0204dd0dc48da1a83d016c3daf4693590c33