msm: bam_dmux: make rx ring buffer pool size configurable
The recieve ring buffer pool size defaults to 32 buffers, which maintains
a good balance between performance and memory consumption.  The hardware
is capable of supporting larger pools, which could be tuned to
non-standard usecases or new requirements.

Add support for an optional dt property that allows the default size to
be overridden with a target specific size.

Change-Id: Ib73b3836d91cfb6ac0c95d2c1f54a01e26a100a2
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>