arm/dt: pm8941: Modify BTM channel index
Update the BTM channel index within the channel
properties to be sequential. We need this sequential
because of the dependency to know which channel needs
to be enabled in the qpnp_adc_tm driver. Having the
BTM channel sequential helps knowing what channel
measurement is being enabled for the corresponding
VADC channel and easier to debug.

Change-Id: Ic1b12210396c8b3e0f962ac7efc46295e7e01996
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>