msm: smd: always leave 4 bytes in SMD FIFO
SMD uses tail == head as FIFO empty and (tail - head) == 1 as FIFO
full, but non-Linux implementations use (tail - head) == 4 as the FIFO
full signal.  This inconsistency can cause a data stall for large packets
because the remote side will not detect (tail - head) == 1 as a
full-FIFO condition.

Change the code to always reserve 4 bytes for the FIFO-full condition.

CRs-Fixed: 471277
Change-Id: Ibc2751a0352c2b6720ece84b5865d44d42273287
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>