mmc: block: Fix issue with block device operation mmc_blk_ioctl
The MMC block device ioctl path is not updated with necessary
mmc_rpm_hold and mmc_rpm_release APIs due to which the device
is either not getting resumed if it is already runtime suspended
or it is getting into runtime suspend state while the ioctl
request is still in progress. This issue would be observed only
if MMC_CAP2_CORE_RUNTIME_PM capability is defined. Fix it by
calling mmc_rpm_hold before start of ioctl and mmc_rpm_release
at the end of ioctl.

Change-Id: If1ccebcde4f797722989963a7c69ce1cfb6484fd
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>