ASoC: msm: qdsp6v2: Proxy Port fixes for wfd session.
When a running WFD session is paused/disconnected, there is
a delay of about 10-15 seconds in switching audio from WFD
sink to target. pcm_write is blocked because pausing/disconnecting
the WFD session closes the backend DAI before a new link is
established. Avoid closing afe if front end PCM session is still
active. When playback is paused and proxy device disabled, afe is
not closed. This results in failure of proxy port open in the second
iteration. Fix the same by closing and opening the proxy.

Change-Id: I32f29da76c10d0fa9c9a7ee89ff07bc38d037256
CRs-Fixed: 472076
Signed-off-by: Krishnankutty Kolathappilly <kkolat@codeaurora.org>