msm: camera: Attach all ISP context during ISP open
When one ISP is active and another ISP opens, the iommu
attach context call will halt the iommu in order to
program attach the context for the second ISP. This will
cause the first ISP write operation to pause and results
in overflow. This change workaround the issue by
attaching both contexts when the first ISP is opened.

Change-Id: I04202c4bbebc250c5f3af72f8800cc581727a9ea
Signed-off-by: Kevin Chan <ktchan@codeaurora.org>