arm/dt: msm_sdcc: update drive strengths
On platforms that support HS200 mode for eMMC and SDR104 mode for
SD cards, the clock drive strength should not be set to the maximum.
If set to maximum drive strength, that would cause the clock signal
to overshoot and not meet the eMMC or SD card specification.

But since current settings used on 8974 have been stable,
overriding the recommended values with them.

CRs-Fixed: 467952
Change-Id: If23f9257856f36e51d71a3149eda9bba376ef37c
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>