ehci: hsic: add support for inactivity_clk
The inactivity clock for hsic is a clock that required for
enabling the hsic bam inactivity timer feature.

The inactivity timer feature is inactivity notification that
supported by the hsic bam hardware and trigger an interrupt
when no data is pulled or pushed on any of the bam pipes,
in both directions for some predefine duration.

Change-Id: I6ce1370a927b2a39534d73db917518ecbc56089a
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>