msm: clock-8226/8610: Remove the XO parent for the q6ss_xo_clk
Peripheral loader code enables certain clocks and leaves
them enabled for as long as the LPASS subsystem is up.
Setting the CXO as the parent for one of these clocks, the
q6ss_xo_clk, will result in XO shutdown being blocked.
Remove CXO as a parent for this clock.

There is no use of this parent relationship as the LPASS
is supposed to vote for CXO soon as it boots.

Change-Id: Ib6108c94a8f835ba558dec7e7407dd4b72d50024
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>