iommu: msm: Ensure power is on in pg fault handler.
There is a race condition between turning off the power to the IOMMU
and accessing registers in the page fault handler. The iommu detach
function might get called just after a page fault happens. The detach
function will turn off the regulators to the IOMMU. Since the page
fault handler is a threaded irq it might not run until after the
detach call has turned off the regulators.

Add a check in the page fault handler that power is still on before
accessing any IOMMU registers.

CRs-fixed: 480439
Change-Id: I3a4a5f4cb7bfa5e1418b68b1f9683df75e0f8038
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>