msm: vidc: Report close done event to client.
This a part of error handling when firmware fails to
respond to close done command. In this case driver times-out
waiting the response. Driver reports this error through the
ioctl. Since, client also expects CLOSE_DONE event in response,
driver should send that event in case of error for proper cleanup
of client. Poll thread in client is exited on receipt of this event.

Change-Id: Ia40903e1b23f514de9adef550aadea04571b8d1a
Signed-off-by: Vinay Kalia <vkalia@codeaurora.org>