msm: mdss: unstage target mixer stage when replacing pipe
If a new pipe comes in with mixer stage that is already taken, the pipe
that is currently holding that mixer stage should first be unstaged.
This ensures that only the pipe taking its place can be placed at that
mixer stage, and avoid situations where error in programming may lead to
previous pipe still being staged with incorrect parameters.

CRs-Fixed: 481364
Change-Id: Ie6d3471c610833a9546e52bc5f3816ddb5cb4101
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>