power: qpnp-bms: Do not use ADC sync API when battery is full
There is a hardware bug that causes the ADC synchronous voltage and
current reads to give garbage current values when the battery FET is
open.

Since the batfet can open and close in microseconds, simply do not
use the ADC synchronous API when the battery is full. Instead, read
battery current from the BMS and read battery voltage from VADC.

CRs-Fixed: 479213
Change-Id: Ic3cc9adbb5d759eb5192e76514bdf55188aa5526
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>