arm/dt: msm8610: default to svs for cx/mx
Vote from regulator is not propagated to RPM because
the notification system does not receive the sleep vote
when the sleep vote matches the current active vote.
As a result super turbo is sent as the voltage
during sleep by default. As a workaround default
to svs for cx and mx temporarily.

Change-Id: I81991b711091e12ebef264b6d7a65eb74f2d1fe4
Signed-off-by: Priyanka Mathur <pmathur@codeaurora.org>