media: dvb: Fix behavior of blocking read operation
When filter is opened in blocking mode and read is requested
while there's no data in filter's output buffer, read operation
is blocked waiting for data events, this was done while mutex
of demux filter is held, causing all ioctl calls to demux
filter to be blocked until read is released (if ever).

This change releases the mutex before going to sleep, once
awakened, mutex is re-obtained before checking the filter's
status.

Change-Id: Iaa9bfbf4672147a6f7872e2e8264ea04c33be79f
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>