power_supply: Add CHARGE_COUNTER_SHADOW property to power supply
Add the "CHARGE_COUNTER_SHADOW" property to the power supply
framework that signifies a shadow charge counter of the power supply.
This should be a shadowed version of the charge counter that is not
cleared automatically.
The units for this property should be in microamp-hours.

Change-Id: Ia9504a3c36af305e42d0503cf791594449249fa0
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>
(cherry picked from commit 4090f4f77cc988d437e3d32a6a4a64749c9df794)