msm: net: ecm: change resource manager feature enable/disable scope
The driver support a feature for enabling and disabling the
resource manager requests that are made prior to data send.
In order to allow this feature to work without the need to restart
the driver (the creation of a resource is done on initialization) the
resource will always be created/destroy regardless of the fact that
the user chooses to enable/disable the resource requests.

Change-Id: I808f87281557a795530598c429ff843fcb388af7
Signed-off-by: Talel Atias <tatias@codeaurora.org>