ASoC: wcd93xx: Fix the system crash caused by virtual port
By design, the same virtual port can only be routed after it is derouted.
There are cases that application can mistakenly routed the same virtual
port which causes the system crash because the link-list to maintain the
list of slimbus port is trashed. The solution is not to add the same port
in the list if it is routed.

Change-Id: I999c2899ca1e3e6f9f6d10f064a4b73f123af047
CRs-Fixed: 480376
CRs-Fixed: 461459
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>