mdss: hdmi: Enable general control packet transmission to support AVMUTE
When the source and the sink are not in sync with respect to their HDCP
cipher and frame counts, it could result in snowframes or similar
artifacts being displayed on the HDMI sink. This can be avoided by
setting AVMUTE before starting HDCP authentication and clearing AVMUTE
after HDCP authentication succeeds. Whenever AVMUTE is set, and if the
sink supports AVMUTE, then the HDCP cipher state and the frame count
are not advanced until after the first encrypted frame is sent after
clearing AVMUTE. This patch implements this functionality by enabling
the transsmission of the general control packet which is used to set
and clear AVMUTE.

CRs-Fixed: 484366
Change-Id: Iaf437611e8b9f7d6cbe94184477f6d72f7af6ad8
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>