msm: smd: remove incorrect use of smd_private.h
smd_private.h is an internal header file used to share private internal
details of the SMD driver to related components for the purposes of
debugging, segregating functionality, etc.  Nothing in smd_private.h
is intended to be exposed for client use, and including it in a file
should be a special exception, not a normal usecase.

Remove client structures from smd_private.h that do not conform to the
purpose of smd_private.h.

Change-Id: Ic7db986d83622cf4300fbc95ff31f129d0ed5f44
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>