msm: camera: Increase AXI and stats CFG timeout
AXI and stats configuration can take up to two frames to be configured
and configuration to take in account. This is because the configuration
happens on the next SOF and is known for sure it will be valid at the
end of the frame.

As the above explains, the need to have at least two frame periods for
each timeout related with the AXI and stats configuration, the logical
safe value for a timeout will be 3 frame periods. Having the assumption
for a maximum of 1000ms for each frame (1 frame per second), this patch
is changing the value to 3000ms for each timeout. This should be safe
enough even for cases when the frame period is getting to higher values
of up to 1500ms (0.66 frames per second).

The previous timeout of 500ms was not enough in the cases when the
sensor is configured to run at 4fps or lower, and there are situations
where the configuration will be falsely assumed to not happen, due to
the low timeout, which gets triggered.

Also as there are already a few places having the same timeout,
calculated on the same base this value is moved to a common define.

Change-Id: I70a65dde3cdcab0abfabb385cb06e11431c1c233
Signed-off-by: Vladislav Hristov <vhrist@codeaurora.org>