arm/dt: msm8226: Add active-set only and sleep-set only regulators for L8
LDO 8 powers the ACPU clock PLL on MSM8226. However, it is currently
configured as active+sleep. Since the sleep request is not
different, rpm-regulator-smd does not send it.

Fix this by adding active-set only and sleep-set only regulator
nodes for L8. Change the consumer mapping for the ACPU PLL supply such
that the new active-set only node is used instead of the old
active+sleep node.

Change-Id: Ie2d3d502b690545ce76ee1ee6720864039440b1a
Signed-off-by: David Keitel <dkeitel@codeaurora.org>