mmc: msm_sdcc: fix pwrsave bit handling
SDCC controller provides the PWRSAVE control bit to automatically
disable the clock to card when there is no activity with card.
SDCC driver provides the compile time option to disable this pwrsave
feature completely but some of the recent driver changes ignore this
and may enable the pwrsave feature even when its not required. This
change ensures that we always check if the pwrsave is allowed or not
before really enabling it.

CRs-Fixed: 490017
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>
(cherry picked from commit 5e26b136842f196741b3d799da982e2ab23e764e)

Change-Id: I3d3629ec3bb8f3ddfcf0bd83423073c72252aa8c
Signed-off-by: Sridhar Nelloru <nellos@codeaurora.org>