input: touch_synaptics_ds5: use I2C RESET COMMAND if possible
Only GPIO RESET was used to reset the touch controller when
there is an error in the touch operation, but controlling
RESET GPIO makes the LCD tearing effect.

Changed to use GPIO RESET only if the I2C command reset is failed.
Because I2C command reset doesn't make the LCD tearing effect,
this will lower the frequency of the LCD tearing issue.

Change-Id: Id0f528f069e1363f2c7da4b6ca22f0b80aad2dfe
Signed-off-by: Jongrak Kwon <jongrak.kwon@lge.com>