ASoC: msm8x10-wcd: Remove CP out of reset state
During charge pump tear-down sequence, the charge pump reset
bit should be set, then clear the bit to keep the codec in
correct state. There is  the bug in the codec driver which
set the CP reset bit twice. Due to this, the charge pump
is never pulled out of reset which causes the PAs to be
in class AB mode instead of class G. Update the driver
to fix the issue.

Change-Id: I344e4a3575c94ffa03c98fec1c75460876023443
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>