mmc: sdhci-msm: fix pwrsave bit handling
SDCC controller provides the PWRSAVE control bit to automatically
disable the clock to card when there is no activity with card.
During the SDCC DLL tuning, PWRSAVE is disabled to make sure that clock
is always running but once the DLL tuning is completed, currently we
enable the PWRSAVE unconditionally irrespective of its previous state.

This change ensures that we always check if the previous state of pwrsave
before really enabling it.

Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>
(cherry picked from commit b677a46bcc4349cfdf2bafb30920f9500118700d)

Change-Id: Iaf21bb4636c98777a5d8b3cb40774eb6a65589da
Signed-off-by: Sridhar Nelloru <nellos@codeaurora.org>