ASoC: msm: qdsp6v2: fix improper usage of atomic operations
There are improper usages of atomic operations that call read decreased
counter with atomic_dec() and atomic_read() while decreament and read
should be atomic.  This is not correct as there is possibility of
preemption between two of operations.
Use atomic_sub_return and coresponding macro to be whole operation is
atomic.

Change-Id: I358bc74c7454c0fd193f284d5358198f14d1d70e
Signed-off-by: Joonwoo Park <joonwoop@codeaurora.org>