usb: usb_bam: hsic consumer request not necessarily resume the hsic core
When hsic consumer is requested, the hsic bam driver vote for resume
the hsic core. In this case hsic core resume is activated and we have
a flag (hsic_host_dev_resumed_from_cons_request) to know if resume
happened in this scenario or not.

However, if hsic core was already resumed when the hsic consumer request
arrived, for example since other driver voted for resume before that, then
no resume will happen and therefore the flag will remain set to true.
Therefore, on the next resume, we will consider the resume as resume
after consumer request and we will prevent from doing necessary operations
such as calling to ipa_resume. Therefore when the consumer is released
we will set this flag back to false so we will know that the next resume
that will happen is since of other resume scenario.

Change-Id: Ied21c2d6c7b299a79bf43a093440cab8123a0407
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>