ASoC: 8226: remove unsed clock API
MSM8226 uses DIV_CLK1 from PMIC 8226 to supply master clock
to audio codec. This clock is set to freq of 9.6MHZ by
default and cannot be set to any other frequency as it is
exposed using RPM.
The function clk_set_rate, if used, will always fail and
will return -1 error. Remove the function to avoid any kind
of confusion and as it is redundant in terms of functionality.
For clocks exposed by RPM, clk_prepare_enable and
clk_disable_unprepare are enough.

Change-Id: I98346d17b878540cb5c4a47531e403393df1b80a
Signed-off-by: Banajit Goswami <bgoswa@codeaurora.org>