power: qpnp-charger: resume charging based on BMS SoC
The charger is designed to report the battery's state of charge (SoC) from
an external driver. This driver could be pmic's internal battery
management system (BMS) driver or a third party chip driver.

Currently the charger driver only resumes charging based on the battery
voltage i.e. it will resume charging on a full battery after the battery
voltage falls below a certain threshold.

Change this so that if the battery's SoC drops below a set
limit, immediately resume charging to keep the battery at full.

Change-Id: Ie60cb2e02df429f62cbca5a3feb35f9d328c3aac
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>