scsi: ufs: Fix device and host reset methods
As of now SCSI initiated error handling is broken because,
the reset APIs don't try to bring back the device initialized and
ready for further transfers.

In case of timeouts, the scsi error handler takes care of handling aborts
and resets. Improve the error handling in such scenario by resetting the
device and host and re-initializing them in proper manner.

Change-Id: Iced6bb8c4db82d77b19ac504b9d52b270bd58ad3
Signed-off-by: Sujit Reddy Thumma <sthumma@codeaurora.org>