msm: camera: Reset core when HALT timeout
CCI has 2 queues to perform i2c transaction on 2 independent i2c
bus. During CCI init, both masters are RESET at the same time.
If some i2c transaction fails, following recovery procedure is
performed on that master.
1) Issue HALT
2) Wait for HALT ACK
3) Issue RESET for that MASTER
4) Wait for RESET ACK

If HALT ACK is not received. Hence step 3 and 4 is not performed.
This causes subsequent transaction on that MASTER to fail in special
case - Front camera transaction timeout while back camera is running
in the background. To avoid this, RESET core even though HALT ACK
is not received.

Change-Id: I5ff634cafef0060794615ef14d29f235230bd2ac
Signed-off-by: Sreesudhan Ramakrish Ramkumar <srramku@codeaurora.org>
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>