ASoC: msm: Fix underrun in LPA driver
Update LPA platform driver to register restart callback.
The implementation of the callback is to issue a pending buffer
written by the application to restart the session when the dequeue
process is stopped due to underrun issue.
The callback is triggered from ALSA core when the framework
detects that the renderer stopped pulling data but the application
is still writing data. It runs in atomic context.

CRs-Fixed: 491077 490992
Change-Id: I1ed35c7d1ae0646fd09aa9a6996c030b3d54c102
Signed-off-by: Krishnankutty Kolathappilly <kkolat@codeaurora.org>