coresight: add jtag fuse state query
Add support to query JTag fuse state and fail the probe if any of
the required fuse(s) are disabled. This enables a single image
with JTag save-restore drivers compiled in to be run on both
Hardware that has JTag functionality disabled via fuse or left
enabled for use.

CRs-Fixed: 497286
Change-Id: I18924b3f4f43fc5601c33a0ce5bf8fba40e9f0dd
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>