msm: smem: Add non-locking smem-get function
For SSR use cases, the failed subsystem may have crashed with the remote
spinlock locked.  In this case, if smem_get_entry() is called before
SMEM has processed the subsystem restart, then it will cause a deadlock.

Add new function that does not acquire the remote spinlock for the SSR
restart use case.

Change-Id: Ieab390855a5bdf32beb1b11884846afef6dbbf98
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>