msm: clock-8974: Add BCR registers for GDSC-collapsible cores
The GDSC driver allows collapse of cores to be disabled
via device tree. Client drivers may still expect the core
to be reset to a clean state after calls to the GDSC APIs,
however.

Specify .bcr_reg addresses for GDSC-collapsible cores so
that the GDSC driver can assert resets to these cores in
lieu of collapsing them. This is similar to what's already
been done for venus0_vcodec0_clk for the same reason.

Change-Id: I1dfdf9562dd49eee7b38a204a063d6787fc4ce85
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>