msm: vidc: Hardcode the timescale for VUI
The timescale is an arbitrary value provided to the firmware to generate
VUI with.  Providing a small timescale is triggering some integer
division inaccuracies in the firmware.  As such provide a nicer
timescale that the f/w can work with.

Change-Id: I6e864a468865cc26564b3ee8fd5ee723de0462ed
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
(cherry picked from commit e681db4520f06726ef00031ac0fa2cf2e5717176)