msm: wfd: Allow client to provide a tolerance for an acceptable FPS
In cases the driver's frameskipping is unacceptable, provide a tuning
parameter that controls the skip frequency.  fpsvariance is the
percentage for the fps that the driver can overshoot by, which allows
for the driver to skip fewer frames.  fpsvariance == 0 forces the driver
to strictly observe the fps, whereas fpsvariance == 100, allows the
driver to completely ignore the fps, allowing to run in a free-running
mode.

Change-Id: Ic25da25807008416246287e2d804d9bd64165320
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>