power: qpnp-charger: fix get_prop_batt_status logic
When battery temperature is cold or hot, charger will
be disabled by HW, the LED which indicates charger
should be off in this condition. Currently if battery
temperature is not OK, BATFET will close but not clear
FAST_CHG_ON_RT_STS. Fix this by checking the BATFET
status in get_prop_batt_status() to report DISCHARGING
in this condition.

CRs-Fixed: 510394
Change-Id: I0810ff412ded4ad60f50d429f0dc54d541c2062d
Signed-off-by: Zhenhua Huang <zhenhuah@codeaurora.org>