diag: Fix incorrect handling of diag polling command
During SSR, diag driver clears its command registration table and
rescans the table for polling command. It sets a flag after taking
into consideration the table entries that are zeroed out leading
to an incorrect registration status. This patch ignores the table
entries that are zeroed out and compares against the correct
polling commands to set the flag.

Change-Id: Ida2b2ee4f3ec03ae17c800ef64d19c2bf2732a79
Signed-off-by: Ravi Aravamudhan <aravamud@codeaurora.org>