arm: dts: msm: Separate power management DTS files for 8974 pro
On 8974 pro targets, the SPM sequences for Core's power collapse modes
will not initiate a handshake with RPM. RPM will switch to Apps sleep
set based on handshake with L2 SPM only. This would reduce the number of
interrupts sent across to RPM during Core power collapse, thereby
reducing the load on RPM.

Also, removed the lpm-level corresponding to Core SPC with L2 GDHS. With
the above mentioned change, RPM would switch to Apps sleep set and could
turn off the HFPLL regulator. If the HFPLL is not turned disabled in
this mode, it could leave the HFPLL in a undefined state.

Change-Id: I873f7e4e1a4c11233a5c480ad4ab7a1e2866efd1
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>