msm: kgsl: Do not call dispatcher fault handler if already called
Do not call the dispatcher fault handler if it has already been
called in the dispatcher work function. This can cause a cmdbatch
to be processed twice in dispatcher if a false fault occurs and
will run the policy on cmdbatch twice without letting the first
policy retire the cmdbatch.

Change-Id: Ie67852a2a63db1c6180f6ec2c1c968f1d416462f
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>