msm: kgsl: Don't pause the dispatcher in adreno_dispatcher_stop
adreno_dispatcher_stop is called whenever the GPU goes into power
collapse.  The next access to the hardware will turn back on the
power (and call adreno_dispatcher_start) and things will progress
normally.  The problem is that in many cases the next intended
hardware access will come from the dispatcher itself after a
context submits a new command batch.  If we put the dispatcher
into PAUSE in adreno_dispatcher_stop() then no new commands
will be submitted until somebody else turns back on the hardware
which may or may not happen.

Don't go into PAUSE in adreno_dispatcher_stop. It is understood
that the next command submission might turn back on the hardware.
Note that we do still go into PAUSE for suspend/resume cases
because in that situation we do want the dispatcher to stop until
we are told otherwise by the system but that is a special case.

Change-Id: Ic0dedbadaa6920908ecf5287ea7ab7d0fd46c127
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>