power: qpnp-charger: add ext ovp support
Some platform configurations require an
external OVP when using a DCP charger and
fast charging. This external OVP is connected
in parallel to reduce the input resistance
of the charge path to allow more current
to be pumped into the battery.

This additional external ovp needs to be
communicated to userspace services since
the input current limit only impacts the
USBIN charge path and not the newly added
path added by the external OVP. To allow for
userspace binaries to adjust the input current
limit correctly also expose the usb trim register
via the power supply framework.

CRs-Fixed: 532962
Change-Id: I4734f4734f89d8c26508e543398b4b291f1bed29
Signed-off-by: David Keitel <dkeitel@codeaurora.org>