diag: Increase the maximum number of SSIDs per msg mask range
Diag was using a max ssid range of 100. This has increased due
to addition of new SSIDs. This patch increases the range to 200
and adds checks to limit the SSID masks to this range. This patch
changes the order of sending the feature masks, sending it before
the msg masks to let the peripherals know the features Apps
support.

Change-Id: Id913712b29208054ab036743ffc4122699501bc3
CRs-Fixed: 530659
Signed-off-by: Ravi Aravamudhan <aravamud@codeaurora.org>