msm: kgsl: Add an helper function to clear the GPU fault after reset
Commit 3f44c3835ed073d02fa08b99a4457c6fadf4606d allowed for multiple
GPU faults to be set at the same time, but in doing so it changed
the behavior of adreno_set_gpu_fault() so that one can no longer
just pass in 0 to clear the fault status after GPU reset.  As a
result faults that occured after the inital fault that triggered
recovery would stay recorded in the driver and cause the GPU
start functions to fail.

Add adreno_clear_gpu_fault() to explicitly clear the fault after
GPU reset.

Change-Id: Ic0dedbad5fcd64f5d334a68731e40cc0dd86b79c
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>