Revert "defconfig: 8226: Set ROW as the default block I/O scheduler"
This reverts commit 39db85c684d7c4a7dd8ea375c29a5ed502e98898.
The reverted commit was identified as the cause of the FS corruptions
observed on 8x26. The type of issues vary, but ultimately the symptom
is device is stuck in android screen internal reason could be file
system is mounted as RO or databases corruption or EROFS error or
EIO errors.

It's reverted till further annalists of the root cause of FS error.

Change-Id: I4d5d653abb5d9652f1baba560d8b444b739ab736
CRs-fixed: 581363
Signed-off-by: Tatyana Brokhman <tlinder@codeaurora.org>