power: qpnp-charger: take into account charge path priority
On the SMBB peripheral there is a possibility of having
both the USB and DC input being connected at the same time.
In this situation there is a requirement to be able to communicate
to userspace services that the higher priority input was
inserted so appropriate action can be taken.

This is achieved as follows:

1. If DC has the higher priority, the USB input
current limit will be reduced to 100 mA to avoid
charging from the USB path.

2. If USB has priority the DC input current
limit will be reduced to 100 mA to avoid charging
from the DC path.

3. Upon the removal of the priority charge path
the secondary charge path input current limit
will be restored to allow charging.

Change-Id: Ib2ad561ef414eca98f1b3a003aaa0eea1fec8269
CRs-Fixed: 637169
Signed-off-by: David Keitel <dkeitel@codeaurora.org>
Signed-off-by: Gopal G Goberu <ggober@codeaurora.org>