msm: kgsl: Add new API's to track device mutex owner.
Add new API's kgsl_mutex_lock/unlock to track device mutex owner.
Mutex owner is an atomic variable and all operations set/clear on
it are also atomic to avoid any race conditions.

If a thread tries to acquire the mutex using new API, try to acquire
the mutex only if mutex owner is not same as current thread. This is
to prevent deadlocks which can happen if a thread tries to acquire
mutex with mutex held.

Change-Id: I743c95e1b86e7994a5a4d87727a8eb122fc05677
Signed-off-by: Tarun Karra <tkarra@codeaurora.org>
Signed-off-by: Abhishek Ranjan <aranja@codeaurora.org>