msm: mdss: Fix stop command mode panel timeout issue
When stop command mode panel, if the panel clock is disabled
already, don't wait for the stop com completion to avoid
wait timeout.
And the vsync handler only can be removed when it is enabled.
Otherwise rdptr_enabled will be reset to VSYNC_EXPIRE_TICK
which will cause IRQ_PING_PONG_RD_PTR can't be enabled when
the vsync handler is added. Then it will cause stop command
mode panel timeout as well.

Change-Id: I8406a552a3fa022a62827f0f78c930139e1eaffd
Signed-off-by: Jilai Wang <jilaiw@codeaurora.org>