ASoC: wcd9320: Adjust the TxFE clock ratio for MAD
If MAD is connected to ADC, the settling of the TxFE should be minimal
to allow MAD to perform successful detection of audio samples. If the
TxFE settling in higher, then the performace of MAD will decrease. Fix
this issue by changing the TxFE CLKDIV ratio when MAD is enabled.

Change-Id: If3e3249fcbf1ea9307974ab5c69baca0add6c005
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>