msm: mdss: update the pp status flags for layermixer
Post Processing(PP) driver caches the status of the features enabled.
Information is used to turn on the PP features when device is resumed
from suspend state. Currently PP status information is not updated for
layer mixer path. Add support to track the layer mixer status
and prevent programming the DSPP blocks if dirty flag is not set.

Change-Id: If3bba46d3d2286385ba7fa1f0221356d436746ef
Signed-off-by: Gopikrishnaiah Anandan <agopik@codeaurora.org>
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>